adding listen trait
